% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{add_GO_data}
\alias{add_GO_data}
\title{Add GO annotations}
\usage{
add_GO_data(df, map_id = "Entry", GO_type = "molecular_function",
  organism = "mouse", slim = FALSE, updateProgress = NULL)
}
\arguments{
\item{df}{a data.frame with protein IDs in column \code{map_id}}

\item{map_id}{column name used to map protein identifiers}

\item{GO_type}{type of GO term ("molecular_function", "biological_process" or "cellular_component")}

\item{organism}{organism for which the annotations have to be mapped}

\item{slim}{logical, use GO_slim annotations}

\item{updateProgress}{logical, function to show progress in shiny app}
}
\value{
a data.frame with an additional GO annotation column
}
\description{
Add GO annotations corresponding to a set of protein identifiers
}
