% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{add_Hallmark_data}
\alias{add_Hallmark_data}
\title{Add Hallmark annotations}
\usage{
add_Hallmark_data(df, map_id = "Gene.names...primary..",
  updateProgress = NULL)
}
\arguments{
\item{df}{a data.frame with protein IDs in column \code{map_id}}

\item{map_id}{column name used to map protein identifiers}

\item{updateProgress}{logical, function to show progress in shiny app}
}
\value{
a data.frame with an additional Hallmark annotation column
}
\description{
Add Hallmark annotations corresponding to a set of protein identifiers
}
