% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{add_KEGG_data}
\alias{add_KEGG_data}
\title{Add KEGG pathway annotations}
\usage{
add_KEGG_data(df, map_id = "Cross.reference..KEGG.",
  organism = "mouse", updateProgress = NULL)
}
\arguments{
\item{df}{a data.frame with KEGG IDs in column \code{map_id}}

\item{map_id}{column name used to map KEGG IDs}

\item{organism}{organism for which the annotations have to be mapped}

\item{updateProgress}{logical, function to show progress in shiny app}
}
\value{
a data.frame with an additional KEGG pathway annotation column
}
\description{
Add KEGG pathway annotations corresponding to a set of KEGG IDs
}
