% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{annotation_enrichment_analysis}
\alias{annotation_enrichment_analysis}
\title{Perform enrichment analysis}
\usage{
annotation_enrichment_analysis(df, sep = NULL, idx_detect,
  annotation_selected = c("Keywords", "Protein.families"),
  names = df$Gene.names...primary.., organism = "mouse",
  two_sided = FALSE, updateProgress = NULL, showProgress = TRUE,
  orderOutput = TRUE)
}
\arguments{
\item{df}{a data.frame with annotations corresponding to each row. Types of annotations are organized by columns. 
For a given type of annotations, annotations are separated by \code{sep}.}

\item{sep}{Character string separating different annotations of a given type}

\item{idx_detect}{indexes of the foreground set.}

\item{annotation_selected}{set of annotations on which to perform the analysis}

\item{names}{row names. Used in the output data.frame}

\item{organism}{organism for which the analysis is to be performed ("mouse" or "human")}

\item{two_sided}{logical, perform a two-sided hypergeometric test}

\item{updateProgress}{logical, function to show progress in shiny app}

\item{showProgress}{logical, show progress in console}

\item{orderOutput}{logical, order annotations by enrichment p-values in the output data.frame}
}
\value{
a data.frame
}
\description{
Perform enrichment analysis using a hypergeometric test for protein annotations stored in a formatted data.frame
}
