% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{append_PPI}
\alias{append_PPI}
\title{Append protein-protein interaction}
\usage{
append_PPI(res, mapping = "names", df_summary = NULL)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{mapping}{name of the \code{InteRactome}'s variable containing gene names}

\item{df_summary}{data.frame with PPI information obatined from a call to \code{create_summary_table_PPI()}}
}
\value{
an \code{InteRactome}
}
\description{
Append protein-protein interaction information to an \code{InteRactome}.
PPI are retrieved from databases IntAct, MINT, BioGRID and HPRD
}
