% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{create_summary_table_PPI}
\alias{create_summary_table_PPI}
\title{Retrieve protein-protein interaction information from databses 
IntAct, MINT, BioGRID and HPRD}
\usage{
create_summary_table_PPI(gene_name)
}
\arguments{
\item{gene_name}{the gene name for which to retrieve PPI}
}
\value{
a data.frame PPI information
}
\description{
Retrieve protein-protein interaction information from databses 
IntAct, MINT, BioGRID and HPRD
}
