% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{filter_annotation_results}
\alias{filter_annotation_results}
\title{Plot the result of the annotation enrichment analysis}
\usage{
filter_annotation_results(df, p_val_max = 0.05,
  method_adjust_p_val = "fdr", fold_change_min = 2, N_annot_min = 2,
  test_depletion = FALSE)
}
\arguments{
\item{df}{a formatted data.frame obtained by the function \code{annotation_enrichment_analysis()}}

\item{p_val_max}{threshold for the enrichment p-value}

\item{method_adjust_p_val}{method to adjust p-value for multiple comparisons}

\item{fold_change_min}{threshold for the enrichment fold-change}

\item{N_annot_min}{minimum number of elements that are annotated in the foreground set}

\item{test_depletion}{logical, test for annotation depletion as well as enrichment}
}
\value{
a data.frame
}
\description{
Plot the result of the annotation enrichment analysis
}
