% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{get_annotations}
\alias{get_annotations}
\title{Get annotations from uniprot for a set of protein identifiers}
\usage{
get_annotations(data, name_id = "Protein.IDs", split_param = ";",
  organism = "mouse", updateProgress = NULL)
}
\arguments{
\item{data}{a data.frame with protein IDs in column \code{name_id}}

\item{name_id}{column name used to map protein identifiers}

\item{split_param}{split character used to separate different protein IDs}

\item{organism}{organism for which the annotations have to be appended}

\item{updateProgress}{logical, function to show progress in shiny app}
}
\value{
an annotated data.frame
}
\description{
Get annotations from uniprot for a set of protein identifiers.
From a set of IDs, keep the first that correspond to a "reviewed" protein, 
or by default the first ID of the set
name_id : column containing the set of protein identifiers separated by "split_param"
organism = c("mouse", "human")
}
