% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{get_annotations_enrichr}
\alias{get_annotations_enrichr}
\title{Get annotations using enrichR}
\usage{
get_annotations_enrichr(data, name_id = "names",
  dbs = "GO_Biological_Process_2018", append_to_data = TRUE)
}
\arguments{
\item{data}{a vector of gene names or a data.frame with gene names in column \code{name_id}}

\item{name_id}{column name used to map gene names}

\item{dbs}{name of the enrichR database. Use \code{enrichR::listEnrichrDbs()} to see available databases.}

\item{append_to_data}{logical, append annotations as a new column}
}
\value{
an annotated data.frame
}
\description{
Get annotations from an enrichR database for a set of genes.
}
