% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{get_annotations_KEGG}
\alias{get_annotations_KEGG}
\title{Create a data.frame with KEGG pathways corrresponding to a set of UniProt IDs}
\usage{
get_annotations_KEGG(id, sep = ";", org = "mmu")
}
\arguments{
\item{id}{Character vector with UniProt IDs}

\item{sep}{Character separating different protein ids}

\item{org}{Organism (ex : "mmu" for mus musculus, "hsa" for homo sapiens).}
}
\value{
a data.frame
}
\description{
Create a data.frame with KEGG pathways corrresponding to a set of UniProt IDs
}
\examples{
id <- c("P26450", "O00459")
df <- get_annotations_KEGG(id = id)
}
