% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{get_annotations_uniprot}
\alias{get_annotations_uniprot}
\title{Create a data.frame with UniProt annotations corrresponding to a set of UniProt IDs}
\usage{
get_annotations_uniprot(
  id,
  sep = ";",
  columns = c("gene_names", "organism_id", "reviewed", "keyword", "protein_families",
    "go"),
  max_keys = 400,
  updateProgress = NULL,
  show_progress = TRUE
)
}
\arguments{
\item{id}{Character vector with UniProt IDs}

\item{sep}{Character separating different protein ids}

\item{columns}{names of uniprot data columns to retrieve. 
Examples include "accession", keyword", "sequence", "go".}

\item{max_keys}{maximum number of field items submitted}

\item{updateProgress}{used to display progress in shiny apps}

\item{show_progress}{Show progress bar}
}
\value{
a data.frame
}
\description{
Create a data.frame with UniProt annotations corrresponding to a set of UniProt IDs
}
\examples{
id <- c("P26450", "O00459")
df <-  get_annotations_uniprot(id = id)
}
