% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{get_PPI_from_HPRD}
\alias{get_PPI_from_HPRD}
\title{Retrieve protein-protein interaction information from HPRD}
\usage{
get_PPI_from_HPRD(gene_name)
}
\arguments{
\item{gene_name}{the gene name for which to retrieve PPI}
}
\description{
Retrieve protein-protein interaction information from HPRD
}
