% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{get_PPI_from_psicquic}
\alias{get_PPI_from_psicquic}
\title{Retrieve protein-protein interaction information using PSICQUIC}
\usage{
get_PPI_from_psicquic(gene_name, taxon_ID = c(9606, 10090),
  provider = c("IntAct", "MINT", "BioGrid"))
}
\arguments{
\item{gene_name}{the gene name for which to retrieve PPI}

\item{taxon_ID}{taxon ID for which to retrieve PPI}

\item{provider}{database from which to retrieve PPI}
}
\value{
a data.frame PPI information
}
\description{
Retrieve protein-protein interaction information using PSICQUIC
}
