% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{identify_reviewed_proteins_ids}
\alias{identify_reviewed_proteins_ids}
\title{Identify which protein IDs correspond to SwissProt reviewed entries}
\usage{
identify_reviewed_proteins_ids(ids, sep = ";", organism = NULL)
}
\arguments{
\item{ids}{vector of protein ids}

\item{sep}{character separating different protein ids}

\item{organism}{taxon id (10090 for Mus musculus, 9606 for human). If null, the most represented organism for ten first protein ids is chosen.}
}
\value{
a data.frame with columns 'id' and 'reviewed'
}
\description{
Identify which protein IDs correspond to SwissProt reviewed entries
}
\examples{
res <- queryup::get_uniprot_data(query=list("gene_exact"="Pik3r1", "organism_id"="10090"))
identify_reviewed_proteins_ids(res$Entry)
}
