% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{retrieve_all_KEGG_pathways}
\alias{retrieve_all_KEGG_pathways}
\title{Create a data.frame with all KEGG pathways for a given organism}
\usage{
retrieve_all_KEGG_pathways(org = "mmu")
}
\arguments{
\item{org}{Organism (ex : "mmu" for mus musculus, "hsa" for homo sapiens).}
}
\value{
a data.frame
}
\description{
Create a data.frame with all KEGG pathways for a given organism
}
\examples{
df <-  retrieve_all_KEGG_pathways(org="mmu")
head(df)
}
