% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pannot.R
\name{plot_annotation_results}
\alias{plot_annotation_results}
\title{Plot the result of the annotation enrichment analysis}
\usage{
plot_annotation_results(df, var_p_val = "fdr",
  fold_change_max_plot = 4, save_file = NULL, ...)
}
\arguments{
\item{df}{a formatted data.frame obtained by the function \code{annotation_enrichment_analysis()}}

\item{var_p_val}{name of the p-value variable}

\item{fold_change_max_plot}{maximal fold-change displayed}

\item{save_file}{path where the plot will be saved}

\item{...}{parameters passed to function \code{filter_annotation_results}}
}
\value{
a plot
}
\description{
Plot the result of the annotation enrichment analysis
}
